Fix topic of MqttConnector which was composed of topic + '?' + queryString. After disconnecting, the queryString was not valid. It is better to compose the topic with the topic name associated with the queryString to avoid any misunderstanding which fixes the issue of stopping MQTT subscribe after unsubscribing a topic
