+++ source/core/connector/MqttConnector.js
@@ -56,6 +56,7 @@ class MqttConnector extends DataConnector {
         });
         this.interval = -1;
         this.id = `mqtt-connector-${randomUUID()}`;
+        this.mqttProvider = undefined;
     }
 
     initBc() {
@@ -66,13 +67,11 @@ class MqttConnector extends DataConnector {
             }, [data]);
         }
         this.broadcastChannel = new BroadcastChannel(this.id);
-        let queryString;
         this.broadcastChannel.onmessage = (message) => {
             if(message.data.message === 'subscribe') {
-                queryString = message.data.queryString;
-                this.doRequest(message.data.topic, message.data.queryString);
+                this.doRequest(message.data.topic);
             } else if(message.data.message === 'unsubscribe') {
-                this.disconnect(message.data.topic+'?'+queryString);
+                this.disconnect(message.data.topic);
             }
         }
     }
@@ -81,7 +80,7 @@ class MqttConnector extends DataConnector {
         let fullUrl = this.getUrl() ;
 
         // only 1 provider by URL
-        if(!(fullUrl in mqttProviders)) {
+        if(!this.mqttProvider) {
             let options = {
                 reconnectPeriod: this.reconnectTimeout,
                 connectTimeout: 30 * 1000
@@ -94,20 +93,20 @@ class MqttConnector extends DataConnector {
                 };
             }
 
-            mqttProviders[fullUrl] = new MqttProvider({
+            this.mqttProvider = new MqttProvider({
                 endpoint: fullUrl,
                 clientId: randomUUID(),
                 options: options,
                 mqttPrefix: this.properties.mqttPrefix
             });
             console.warn(`Stored MQTT provider into cache: ${fullUrl}`);
-            mqttProviders[fullUrl].connect();
-            mqttProviders[fullUrl].checkStatus = this.checkStatus;
+            this.mqttProvider.connect();
+            this.mqttProvider.checkStatus = this.checkStatus;
             this.checkStatus(Status.CONNECTED);
         } else {
             console.warn(`Getting MQTT provider from cache: ${fullUrl}`);
         }
-        return mqttProviders[fullUrl];
+        return this.mqttProvider;
     }
 
     checkStatus(status) {
@@ -121,7 +120,7 @@ class MqttConnector extends DataConnector {
     doRequest(topic = '',queryString= undefined) {
         const mqttProvider = this.getMqttProvider();
 
-        mqttProvider.subscribe(`${topic}?${queryString}`, this.onMessage).then(() => {
+        mqttProvider.subscribe(topic, this.onMessage).then(() => {
             this.onChangeStatus(Status.CONNECTED);
         });
     }
@@ -140,7 +139,7 @@ class MqttConnector extends DataConnector {
         // this.init = false;
         // this.closed = true;
         // find the client
-        const client = mqttProviders[this.getUrl()];
+        const client = this.mqttProvider;
 
         if (isDefined(client) && client.isConnected()) {
             if(!topic) {
@@ -168,13 +167,13 @@ class MqttConnector extends DataConnector {
     }
 
     isConnected() {
-        return isDefined(mqttProviders[this.getUrl()]) && mqttProviders[this.getUrl()].isConnected();
+        return isDefined(this.mqttProvider) && this.mqttProvider.isConnected();
     }
 
     reset() {
         this.disconnect();
         console.log(`Remove provider from cache: ${this.getUrl()}`)
-        delete mqttProviders[this.getUrl()];
+        this.mqttProvider.reset();
     }
 }
